Refactor: 정렬 기능 커스텀훅으로 분리
