+++ src/hooks/index.ts
@@ -3,5 +3,6 @@ export * from './useChatboxHelper';
 export * from './useCreateRightChart';
 export * from './useDropdown';
 export * from './useKoreanTicker';
+export * from './useTableSort';
 export * from './useTheme';
 export * from './useWidgetTickers';
